requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
Django==3.2.10
gunicorn==20.1.0
idna==3.3
newsapi==0.1.1
newsapi-python==0.2.6
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
stepic==0.5.0
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
